+++ src/main/java/com/back/together02be/stock/client/KisPriceClient.java
@@ -1,9 +1,8 @@
 package com.back.together02be.stock.client;
 
-import com.back.together02be.stock.dto.KisPriceRes;
-import com.back.together02be.stock.dto.KisTokenRes;
+import com.back.together02be.stock.dto.response.KisPriceRes;
+import com.back.together02be.stock.dto.response.KisTokenRes;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestClient;
