+++ src/main/java/com/back/together02be/stock/dto/response/StockListRes.java
@@ -1,4 +1,4 @@
-package com.back.together02be.stock.dto;
+package com.back.together02be.stock.dto.response;
 
 //프론트에 보내기 위해 데이터 형태를 따로 정의
 //DB (Stock) -> 가공 -> API 응답
