+++ src/main/java/com/back/together02be/stock/service/StockService.java
@@ -4,9 +4,9 @@ package com.back.together02be.stock.service;
 import com.back.together02be.infra.kis.KisWebSocketClient;
 import com.back.together02be.stock.cache.StockPriceCache;
 import com.back.together02be.stock.client.KisPriceClient;
-import com.back.together02be.stock.dto.KisPriceRes;
+import com.back.together02be.stock.dto.response.KisPriceRes;
 import com.back.together02be.stock.dto.RealtimeStockPrice;
-import com.back.together02be.stock.dto.StockListRes;
+import com.back.together02be.stock.dto.response.StockListRes;
 import com.back.together02be.stock.dto.response.StockPriceRes;
 import com.back.together02be.stock.entity.Stock;
 import com.back.together02be.stock.repository.StockRepository;
